Refactor TOC module into smaller classes
